likely successful upgrade of odor sequence information
now in addition to having the odor ID which must be decoded, with each
experiment tsne_data.mat file, we have a odor_conc_inf file which
explicitly has the odor sequence and timing. If the odor list changes
(becuase odors or concentrations are added or removed), the odor IDs
will change, but the odor_conc_inf table will remain the same. The odor
ID can be computed directly from this table in combination with the
odor_inf lists, which now should also be saved with each tsne_data.mat
file.